rewrite build file to more idiomatic sbt
